UBUNTU: [Packaging] use olddefconfig in stamp-prepare-tree
You do not need to invoke syncconfig here because Kbuild takes care of
it before staring the build. prepare and scripts are unneeded for the
same reason.

All the use of 'syncconfig' in debian/rules is suspicous because
syncconfig may stop to ask for user's preference when it finds a
new config option. There is nobody to answer the Kconfig's question
when the package is being built in auto-builders such as cbd, launchpad.

Presumably what you want here is olddefconfig, which is a non-interactive
command to set any config option to its default.

Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>
Acked-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>