UBUNTU: [Packaging] debian/rules: Fix PPA debug package builds
Build PPAs have a configuration option 'Build debug symbols'. Currently,
if we build in a PPA that has this option disabled, we still build
the debug packages but we remove them from debian/files so they don't
get published. That seems wrong, we shouldn't even build the packages.

The LP build system adds the option 'noautodgbsym' to the env variable
DEB_BUILD_OPTIONS in such a case [1]. Use that to disable debug package
builds.

While at it, clean up the mangling of debian/files and remove logic
that was added in 2010 to support future LP buildd features that never
materialized [2].

[1] https://bugs.launchpage.net/bugs/1623256
[2] https://git.launchpad.net/~ubuntu-kernel/ubuntu/+source/linux/+git/lucid/commit/?id=b56658a9b08b5114d973403fda3cb6216123ace6

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>