UBUNTU: SAUCE: apparmor4.0.0 [82/87]: apparmor: add additional flags to extended permission.
With the switch to extended permissions the ACCEPT2 became unused
freeing up to store state specific flags. The first flags to be
stored are FLAG_OWNER and FLAG other which paves the way to final
merge file and policydb perms into a single permission table.

Lookups based on the objects ownership conditional will still need
separate fns, for now. But make the function slightly more generic
instead of being dedicated to to file permission.

Signed-off-by: John Johansen <john.johansen@canonical.com>