UBUNTU: SAUCE: apparmor4.0.0 [81/87]: apparmor: convert easy uses of unconfined() to label_mediates()
This begins the transition from doing a broad unconfined() check to a
finer label_mediates() check. Not only does this allow for finer control
but it simplifies some code paths, removing redundant checks, will allow
us to refactor and simplify the code more, and is generally a small
performance win.

Note: capabilities and resource need to be converted to use pdb
      before they can use label_mediates

Signed-off-by: John Johansen <john.johansen@canonical.com>