UBUNTU: [Packaging] re-introduce linux-doc as an empty package
In commit fd8ee93c3a38 ("UBUNTU: [Packaging] drop useless linux-doc") we
decided to remove the linux-doc package, saving in this way a lot of
space required to produce the documentation content (that can be easily
found online).

However, this change broke some user-space packages that are still
requiring linux-doc as a dependency.

Therefore, re-introduce an empty linux-doc package to prevent the
breakgage and at the same time avoid to re-introduce the expensive build
dependencies.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>